Mesy Solution
This should be the most basic way of solving this problem. We go down the tree and at each point we see what the infection rate for the next day is and we choose the smaller of them.  This then works its way up to the beginning tellings us only the smallest number of infections. In the case of a tie we choose to be in los angeles because of DisneyLand